// src/app/api/model/transaction.ts
/**
 * Personal Finance Management API
 *
 * Contact: dragisa@yahoo.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Category } from './category';
import { RecurringTransaction } from './recurringTransaction';


export interface Transaction { 
    id?: number;
    description?: string;
    amount?: number;
    date?: string;
    username?: string;
    type?: Transaction.TypeEnum;
    category?: Category;
    recurringTransactions?: Array<RecurringTransaction>;
    createdAt?: string;
    updatedAt?: string;
}
export namespace Transaction {
    export type TypeEnum = 'INCOME' | 'EXPENSE';
    export const TypeEnum = {
        Income: 'INCOME' as TypeEnum,
        Expense: 'EXPENSE' as TypeEnum
    };
}